table64: preserve inert oversized declarations
Why:
- four valid declaration-only table64 modules carry maxima above the executable 16,384-entry reservation and were rejected despite zero observable growth or export

What:
- admit only one inert unexported local funcref table64 declaration with an allocatable minimum
- retain exact uint64 declared maxima through metadata and codec v26 while deriving a header/minimum-only runtime capacity
- reject malformed values above u64 and keep oversized executable/exported shapes fail-closed
- update table metadata widths without growing Compiled or tableDef

Proof:
- go test ./src/wago -run '^TestStagedInertOversizedTable64DeclarationsPreserveExactMax$' -count=1 -v
- go test ./src/wago ./src/core/compiler/frontend -count=1
- go test ./... -run '^$'
- pinned table64 accounting: 2,802 commands, 105 modules, 2,600 assertions, 2 remaining gates, zero blocked or hidden failures
- BenchmarkStagedTable64Size: 37.01-38.31 ns/op, 0 B/op, 0 allocs/op
- TestExternrefTableStructFootprintsRemainBounded: PASS

Docs:
- unchanged; iteration ledger is reserved for the required fourth documentation commit

+++ src/wago/globals.go
@@ -540,7 +540,7 @@ type ElemInit struct {
 type tableDef struct {
 	ImportKey      string  // non-empty only for imported nonzero table indexes
 	Size           int     // local size, or imported minimum when ImportKey is non-empty
-	Max            int     // local runtime capacity, or imported declared maximum
+	Max            uint64  // exact local declared maximum when HasMax; otherwise runtime reserve; imported maximum when ImportKey is non-empty
 	Type           ValType // zero is the hand-built legacy funcref shape
 	ValueTypeIndex uint32
 	HasValueType   bool
@@ -663,7 +663,7 @@ type Compiled struct {
 	TableValueTypeIndex uint32
 	TableHasValueType   bool
 	TableSize           int        // initial/current table-0 length
-	TableMax            int        // table-0 allocated capacity/max; zero means TableSize for older hand-built metadata
+	TableMax            uint64     // exact declared maximum when TableHasMax; otherwise runtime reserve; zero means TableSize for older hand-built metadata
 	HasTableInitFunc    bool       // table-0 initializer is a non-null ref.func payload
 	TableHasMax         bool       // local table-0 declaration has an explicit maximum
 	TableAddr64         bool       // table-0 indexes and limits use the Core 3 i64 address form
